requirements.txt
streamlit==1.41.1
pandas==2.2.3
python-dotenv==1.0.1
groq==0.13.1
pickle
numpy